Move db/stores/test_utils into root test_utils module
