+++ aquadoggo/src/db/stores/mod.rs
@@ -8,5 +8,3 @@ mod log;
 mod operation;
 mod schema;
 mod task;
-#[cfg(test)]
-pub mod test_utils;
