+++ aquadoggo/src/graphql/client/dynamic_types/tests.rs
@@ -8,7 +8,7 @@ use rstest::rstest;
 use serde_json::json;
 use serial_test::serial;
 
-use crate::db::stores::test_utils::{add_schema, test_db, TestDatabase, TestDatabaseRunner};
+use crate::test_utils::{add_schema, test_db, TestDatabase, TestDatabaseRunner};
 use crate::test_utils::graphql_test_client;
 
 #[rstest]
