+++ aquadoggo/src/graphql/client/static_query.rs
@@ -52,7 +52,7 @@ mod tests {
     use serde_json::json;
     use serial_test::serial;
 
-    use crate::db::stores::test_utils::{test_db, TestDatabase, TestDatabaseRunner};
+    use crate::test_utils::{test_db, TestDatabase, TestDatabaseRunner};
     use crate::test_utils::graphql_test_client;
 
     #[rstest]
