+++ aquadoggo/src/graphql/client/tests.rs
@@ -11,7 +11,7 @@ use rstest::rstest;
 use serde_json::json;
 use serial_test::serial;
 
-use crate::db::stores::test_utils::{
+use crate::test_utils::{
     add_document, add_schema, test_db, TestDatabase, TestDatabaseRunner,
 };
 use crate::test_utils::graphql_test_client;
