+++ aquadoggo/src/graphql/client/utils.rs
@@ -41,7 +41,7 @@ mod tests {
     use p2panda_rs::test_utils::fixtures::random_key_pair;
     use rstest::rstest;
 
-    use crate::db::stores::test_utils::{add_document, test_db, TestDatabase, TestDatabaseRunner};
+    use crate::test_utils::{add_document, test_db, TestDatabase, TestDatabaseRunner};
     use crate::graphql::client::utils::validate_view_matches_schema;
 
     #[rstest]
