+++ aquadoggo/src/graphql/schema.rs
@@ -204,7 +204,7 @@ mod test {
     use serde_json::{json, Value};
     use serial_test::serial;
 
-    use crate::db::stores::test_utils::{add_schema, test_db, TestDatabase, TestDatabaseRunner};
+    use crate::test_utils::{add_schema, test_db, TestDatabase, TestDatabaseRunner};
     use crate::test_utils::graphql_test_client;
 
     #[rstest]
