+++ aquadoggo/src/materializer/tasks/dependency.rs
@@ -184,7 +184,7 @@ mod tests {
     use p2panda_rs::WithId;
     use rstest::rstest;
 
-    use crate::db::stores::test_utils::{
+    use crate::test_utils::{
         add_document, add_schema, doggo_schema, test_db, TestDatabase, TestDatabaseRunner,
     };
     use crate::materializer::tasks::reduce_task;
