+++ aquadoggo/src/test_utils/client.rs
@@ -14,7 +14,7 @@ use tokio::task::{self, JoinHandle};
 use tower::make::Shared;
 use tower_service::Service;
 
-use crate::db::stores::test_utils::TestDatabase;
+use crate::test_utils::TestDatabase;
 use crate::graphql::GraphQLSchemaManager;
 use crate::http::{build_server, HttpServiceContext};
