+++ aquadoggo/src/test_utils/mod.rs
@@ -3,7 +3,11 @@
 mod client;
 mod config;
 mod db;
+mod runner;
+mod helpers;
 
+pub use runner::{test_db, with_db_manager_teardown, TestDatabaseManager, TestDatabaseRunner};
+pub use helpers::{add_document, add_schema, build_document, doggo_fields, doggo_schema};
 pub use client::{TestClient, shutdown_handle, graphql_test_client};
 pub use config::{TestConfiguration, TEST_CONFIG};
-pub use db::{drop_database, initialize_db, initialize_db_with_url};
+pub use db::{drop_database, initialize_db, initialize_db_with_url, TestData, TestDatabase};
